Move DukeException to output package
