+++ src/main/java/output/DukeException.java
@@ -1,4 +1,4 @@
-package exception;
+package output;
 
 public class DukeException extends Exception {
     public DukeException(String e) {
